Restrutcured multiple event schema versions tests
